FIX: 디자인 패턴 리팩토링
- 라우터 폴더 구조 리팩토링
