+++ client/src/issues/issues-list-section.jsx
@@ -35,6 +35,8 @@ const StyledListSortCheckBoxInput = styled.input.attrs({
 `;
 
 const StyledListSortOpenClosedDiv = styled.div`
+    display: flex;
+    align-items: center;
     width: 500px;
     height: auto;
 `;
@@ -54,11 +56,9 @@ const StyledListSortOpenClosedCheckBoxLabel = styled.label`
             radioToString="open";
         }
 
-        console.log(props.htmlFor)
-        console.log(radioToString)
-
         return props.htmlFor === radioToString ? "lightgray" : "black"
     }};
+    margin-left:${props => props.htmlFor === 'closed' ? "3%" : "0%"} ;
 `;
 
 const StyledListSortOptions = styled.div`
@@ -105,14 +105,10 @@ const IssuesListSection = () => {
                 <StyledListSortOpenClosedDiv>
                         <StyledListSortOpenClosedCheckBox onChange={onOpenClosedRadioChange} id="open"/>
                         <StyledListSortOpenClosedCheckBoxLabel htmlFor="open" openClosedRadio={openClosedRadio}>
-                            <svg viewBox="0 0 16 16" version="1.1" width="16" height="16" aria-hidden="true">
-                                <path d="M8 1.5a6.5 6.5 0 100 13 6.5 6.5 0 000-13zM0 8a8 8 0 1116 0A8 8 0 010 8zm9 3a1 1 0 11-2 0 1 1 0 012 0zm-.25-6.25a.75.75 0 00-1.5 0v3.5a.75.75 0 001.5 0v-3.5z"></path>
-                            </svg>
-                            Open
+                            ⓘ Open
                         </StyledListSortOpenClosedCheckBoxLabel>
-
                         <StyledListSortOpenClosedCheckBox onChange={onOpenClosedRadioChange}  id="closed" />
-                        <StyledListSortOpenClosedCheckBoxLabel htmlFor="closed" openClosedRadio={openClosedRadio}>Closed</StyledListSortOpenClosedCheckBoxLabel>
+                        <StyledListSortOpenClosedCheckBoxLabel htmlFor="closed" openClosedRadio={openClosedRadio}>✔ Closed</StyledListSortOpenClosedCheckBoxLabel>
                 </StyledListSortOpenClosedDiv>
                 <StyledListSortOptions>
                     <DropdownMenu
