+++ server/routes/comment.js
@@ -1,69 +1,17 @@
 import express from "express";
-import pool from "../db/connection";
-import query from "../db/query";
-import isLoggedIn from "./auth";
+
+import isLoggedIn from "../middleware/auth";
+import CommentService from "../service/comment-service"
 
 const router = express.Router();
 
-router.post("/getComment", isLoggedIn, async (req, res) => {
-    const issueId = req.body.issueId;
+router.post("/getComment", isLoggedIn, CommentService.getComment);
 
-    const connection = await pool.getConnection();
-    const [rows] = await connection.query(query.getComment, [issueId]);
-    console.log(rows);
-    res.json(rows);
-});
+router.post("/insertcomment", isLoggedIn, CommentService.insertComment);
 
-router.post("/insertcomment", isLoggedIn, async (req, res) => {
-    const userId = req.body.userId;
-    const comment = {
-        issueId: req.body.issueId,
-        writingTime: req.body.writingTime,
-        comment: req.body.comment
-    };
+router.put("/", isLoggedIn, CommentService.updateComment);
 
-    console.log(comment);
-
-    const connection = await pool.getConnection();
-    const [rows1] = await connection.query(query.insertComment, [userId, comment.writingTime, comment.comment]);
-    const [rows2] = await connection.query(query.insertIssueCommentRelation, [comment.issueId, rows1.insertId]);
-
-    if (rows1.affectedRows > 0 && rows2.affectedRows > 0) {
-        res.json({ message: "success" });
-    }
-
-    else {
-        res.json({ messages: "Error" });
-    }
-});
-
-router.put("/", isLoggedIn, async (req, res) => {
-    const comment = {
-        ID: req.body.ID,
-        comment: req.body.comment,
-        writingTime: req.body.writingTime,
-    };
-
-    const connection = await pool.getConnection();
-    const [rows] = await connection.query(query.updateComment, [comment.comment, comment.writingTime, comment.ID]);
-    if(rows.affectedRows > 0) {
-        res.json({message: "success"});
-    }
-    else {
-        res.json({messages: "Error"});
-    }
-});
-
-router.delete("/", isLoggedIn, async (req, res) => {
-    const ID = req.body.ID;
-
-    const connection = await pool.getConnection();
-    const [rows] = await connection.query(query.deleteComment, [ID]);
-
-    if(rows.affectedRows > 0) {
-        res.json({message: "success"});
-    }
-});
+router.delete("/", isLoggedIn, CommentService.deleteComment);
 
 export default router;
