--- server/routes/hello.js
@@ -1,12 +1,0 @@
-import express from "express";
-
-const router = express.Router();
-
-router.post("/", (req, res) => {
-    console.log(req.body);
-    res.json({
-        message: "Hello! I'm from Express Server!",
-    });
-});
-
-export default router;
