+++ server/routes/index.js
@@ -1,24 +1,18 @@
 import express from "express";
-import hello from "./hello";
-import signIn from "./signIn";
-import signUp from "./signUp";
-import signOut from "./signOut";
 import issue from "./issue";
 import comment from "./comment";
 import label from "./label";
 import milestone from "./milestone";
 import user from "./user";
+import oAuth from "./oAuth";
 
 const router = express.Router();
 
-router.use("/hello", hello);
-router.use("/signIn", signIn);
-router.use("/signUp", signUp);
-router.use("/signOut", signOut);
 router.use("/issue", issue);
 router.use("/comment", comment);
 router.use("/label", label);
 router.use("/milestone", milestone);
 router.use("/user", user);
+router.use("/oAuth", oAuth);
 
 export default router;
