+++ server/routes/milestone.js
@@ -1,14 +1,10 @@
 import express from "express";
-import isLoggedIn from "./auth";
-import pool from "../db/connection";
-import query from "../db/query";
+
+import isLoggedIn from "../middleware/auth";
+import MilestoneService from "../service/milestone-service";
 
 const router = express.Router();
 
-router.get("/", isLoggedIn, async(req, res) => {
-    const connection = await pool.getConnection();
-    const [rows] = await connection.query(query.getMilestone);
-    res.json(rows);
-});
+router.get("/", isLoggedIn, MilestoneService.getMilestone);
 
 export default router;
