--- server/service/db.js
@@ -1,3 +1,0 @@
-import db from "mysql2";
-
-export default db;
